Exchange two elements of a list, and answer every index
`swap(list, i, j) -> Bool` in `core.list`: `false` when either index is out of
range, and then nothing moved (user decision). A total exchange, so nothing
enters the list and nothing leaves — which is what lets a list of obligations
have it when it can have no positional write at all. Three of `demo/heap.sv`'s
four errors; the demo is down to one.

`[col-bounds]` was a label `core.list` referenced and no spec defined. Writing it
exposed that the Rust backend broke its own posture for a *literal* index: the
lowerings cast straight to `usize`, a literal takes its type from the cast
target, and `(-1) as usize` is rustc's E0600 — so `get(xs, -1)` type-checked and
then failed to build, while a variable holding `-1` worked. Every index position
now casts through `i64`, which is one `as i64` in every Rust golden and example.
Kotlin never had the bug, so the shared e2e case is the assertion that Rust
agrees index by index.

+++ examples/actors/rust/core/set.rs
@@ -18,7 +18,7 @@ pub struct SetYield<T: Clone + 'static> {
 }
 
 pub fn next__7<T: Clone>(p: &mut SetYield<T>) -> Union2<T, Finished> {
-    let mut elem = p.items.get((p.at) as usize).cloned();
+    let mut elem = p.items.get((p.at) as i64 as usize).cloned();
     if elem.is_none() {
         return Union2::<T, Finished>::U2(finished());
     }
