+++ examples/collections/rust/core/list.rs
@@ -17,7 +17,7 @@ pub fn non_empty_list<T: Clone>(first: &T, rest: Vec<T>) -> Vec<T> {
 }
 
 pub fn first<T: Clone>(list: &Vec<T>) -> &T {
-    return list.get((0) as usize).unwrap();
+    return list.get((0) as i64 as usize).unwrap();
 }
 
 pub fn iter__3<T: Clone>(list: &Vec<T>) -> ListYield<'_, T> {
@@ -31,7 +31,7 @@ pub struct ListYield<'s, T: Clone + 'static> {
 }
 
 pub fn next__5<'s, T: Clone>(p: &mut ListYield<'s, T>) -> Union2<&'s T, Finished> {
-    let mut elem = p.items.get((p.at) as usize);
+    let mut elem = p.items.get((p.at) as i64 as usize);
     if elem.is_none() {
         return Union2::U2(finished());
     }
